move functions into resolve package
