+++ cmd/kritis/kubectl/plugins/resolve/plugin.yaml
@@ -1,3 +1,3 @@
 name: "resolve-tags"
 shortDesc: "resolve images in kubernetes yamls from tags to digests"
-command: "./main"
+command: "./resolve-tags"
